- fixed compile warning with gcc on solaris
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@1404 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/enum/enum.c
@@ -58,7 +58,7 @@ static int cclen(const char *number)
 	d1 = number[0];
 	d2 = number[1];
 	
-	if (!isdigit(d2)) 
+	if (!isdigit((int)d2)) 
 		return(0);
 
 	switch(d1) {
